requirements.txt
django-paypal
stream-django
django-extensions
django-chatter
django-cron
